# zhihuitushu/src/main/java/com/yt/zhihuitushu/dto/request/LoginDTO.java
package com.yt.zhihuitushu.dto.request;

import jakarta.validation.constraints.NotBlank;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class LoginDTO {

    @NotBlank(message = "账号不能为空")
    private String account;

    @NotBlank(message = "密码不能为空")
    private String password;

    private Boolean rememberMe = false;
}