# zhihuitushu/src/main/java/com/yt/zhihuitushu/dto/response/ApiResponse.java
package com.yt.zhihuitushu.dto.response;

import lombok.Getter;

@Getter
public class ApiResponse<T> {

    private final Integer code;
    private final String message;
    private final T data;

    private ApiResponse(Integer code, String message, T data) {
        this.code = code;
        this.message = message;
        this.data = data;
    }

    public static <T> ApiResponse<T> success(T data) {
        return new ApiResponse<>(200, "success", data);
    }

    public static <T> ApiResponse<T> success(String message, T data) {
        return new ApiResponse<>(200, message, data);
    }

    public static <T> ApiResponse<T> error(Integer code, String message) {
        return new ApiResponse<>(code, message, null);
    }

    public static <T> ApiResponse<T> error(String message) {
        return new ApiResponse<>(500, message, null);
    }
}